requirements.txt
djangorestframework
python-decouple
python-dotenv
djangorestframework-simplejwt
psycopg2-binary